Add base setup project
