+++ HomeWork14/Sources/App/SceneDelegate.swift
@@ -1,0 +1,27 @@
+//
+//  SceneDelegate.swift
+//  HomeWork14
+//
+//  Created by Игорь Николаев on 18.01.2023.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+       let viewController = ViewController()
+        let navigationController = UINavigationController(rootViewController: viewController)
+        window?.rootViewController = navigationController
+        window?.makeKeyAndVisible()
+    }
+}
+
+
+
+
