+++ HomeWork14/Sources/ViewController.swift
@@ -1,0 +1,43 @@
+//
+//  ViewController.swift
+//  HomeWork14
+//
+//  Created by Игорь Николаев on 18.01.2023.
+//
+
+import UIKit
+
+
+class ViewController: UIViewController {
+
+    //: MARK: - UI Elements
+
+
+
+
+    //: MARK: - Lifecycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupView()
+        setupHierarchy()
+        setupLayout()
+    }
+
+    //: MARK: - Setups
+
+    private func setupView() {
+
+    }
+
+    private func setupHierarchy() {
+
+    }
+
+    private func setupLayout() {
+        NSLayoutConstraint.activate([
+
+        ])
+    }
+
+}
