--- HomeWork14/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  HomeWork14
-//
-//  Created by Игорь Николаев on 18.01.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
